fix: eliminate all compiler warnings
- Add unused parameter markers (void)param to silence unused parameter warnings
  in: animation.c, browser.c, cfs_scheduler.c, ipc.c, memory.c, slab.c, syscall.c
- Fix sign comparison warnings in browser.c by using int32_t consistently
  for button/tab calculations mixed with unsigned mouse coordinates
- Mark unused static functions with __attribute__((unused)) in:
  fat.c (fat_alloc_cluster, fat_free_cluster_chain)
  net.c (net_checksum_ipv4)
  rtl8139.c (rtl8139_close)
- Mark unused const variable sched_prio_to_wmult with __attribute__((unused))
  in cfs_scheduler.c
- Fix multi-line comment warning in framebuffer.c (backslash escape)
- Fix int-to-pointer-cast warnings in memory_compress.c using uintptr_t
- Fix pointer comparison warnings in slab.c using uintptr_t casts

Result: Clean build with zero warnings (-Wall -Wextra)

+++ src/kernel/memory_compress.c
@@ -209,7 +209,7 @@ int32_t memory_compress_page(uint32_t page_addr) {
     // Try to compress
     uint8_t temp_buffer[4096 * 2]; // Compression might expand slightly
     uint32_t compressed_size = memory_compress(
-        (uint8_t *)page_addr, page_size,
+        (uint8_t *)(uintptr_t)page_addr, page_size,
         temp_buffer, sizeof(temp_buffer)
     );
     
@@ -232,7 +232,7 @@ int32_t memory_compress_page(uint32_t page_addr) {
     block->last_access = timer_ticks();
     
     // Zero out the original page (mark as swapped)
-    memset((void *)page_addr, 0, page_size);
+    memset((void *)(uintptr_t)page_addr, 0, page_size);
     
     return 0;
 }
@@ -246,12 +246,12 @@ int32_t memory_decompress_page(uint32_t original_addr) {
         if (compressed_blocks[i].original_addr == original_addr) {
             compressed_block_t *block = &compressed_blocks[i];
             
-            int32_t result = memory_decompress(
-                &compressed_pool[block->compressed_addr],
-                block->compressed_size,
-                (uint8_t *)original_addr,
-                block->original_size
-            );
+        int32_t result = memory_decompress(
+            &compressed_pool[block->compressed_addr],
+            block->compressed_size,
+            (uint8_t *)(uintptr_t)original_addr,
+            block->original_size
+        );
             
             if (result == 0) {
                 // Remove from compressed pool (simple approach: mark as free)
